requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
configparser==3.5.0       # via flake8
coverage==4.4.1           # via pytest-cov
docutils==0.14
enum34==1.1.6             # via flake8
flake8==3.4.1
freezegun==0.3.9
isort==4.2.15
mccabe==0.6.1             # via flake8
multilint==2.1.0
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pygments==2.2.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.2.2
python-dateutil==2.6.1    # via freezegun
pytz==2017.2
six==1.10.0               # via freezegun, python-dateutil
sqlparse==0.2.3
